requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
gunicorn==23.0.0

click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1

matplotlib==3.10.1
numpy==1.26.4
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
packaging==23.2
google-generativeai==0.8.3